verification

+++ controllers/controller-list-user.php
@@ -1,17 +1,16 @@
 <?php 
-if (!isset($_SESSION['user'])) {
-    header("Location: login.php");
-  exit;  
+if (!isset($_SESSION['user']) || $_SESSION['user']['role_id_role'] != 1) {
+  header("Location: login.php");
+  exit;
 }
 
 
 
 require_once '../config.php';
 require_once '../models/database.php';
-require_once '../models/users.php';
+require_once '../models/user.php';
 
 if(isset($_GET['user_id'])) {
-  echo 'delete patient';
   $attributDel = new Users();
   $userDel = $attributDel->deleteUser($_GET['user_id']);
 }
